move TextInput to inputs directory
